+++ tests/structures_test.py
@@ -317,7 +317,7 @@ def test_inheritance__field_redefinition_crashes():
 def test_load__discarded_fields_not_present():
     class _Dumb(binobj.Struct):
         first = fields.String(size=2)
-        reserved = fields.Bytes(const=b"\0\0", discard=True)
+        reserved = fields.Bytes(const=True, default=b"\0\0", discard=True)
         last = fields.String(size=2)
 
     loaded = _Dumb().from_bytes(b"AB\0\0CD").to_dict()